Fix an off-by-one mistake (#360)
In PR #354, I introduced a mechanism that counts how many times a given
datum is present in the MockChain, and only deletes the entry in the
'mcstDatums' once that count reaches zero. I realized just now that I
wrote the logic in such a way as to delete the entry *after* the count
has reached zero and would be decremented to -1. This means that entries
are never deleted, which defeats the memory saving purpose that of the
whole operation... This commit rectifies this.