Update to cardano api 10.13 alongside CNE (#461)
This is a major update of both Cardano Node Emulator and Cooked Validators. As of this update, CV is up to date with the most recent version of cardano-api and associated libraries. This means that post-Chang scripts are now fully supported and thus scripts can be used for a variety of purposes by default.