Optics revamping (#484)
Optics have been present since cooked V2 but they suffered from various
shortcomings. They were either seen as too complicated and
over-engineered for our use cases, or not developed enough to actually
be useful in practical use cases. They either needed to go, or to be
revamped entirely.

After giving it some thought and actually using them in practice, I came
to the conclusion that they were indeed a powerful tool that could help
make cooked-validators more consistent and to-the-point, and thus needed
a full refactor. This is the goal of this PR, which introduces several
changes:

* A documentation file is now available on optics in cooked-validators
* Optics are now complete, that is everything that can be focused on has
an optic defined
* Wherever optics can be used, they are
* Optics are all "1st degree", that is we don't provide optics that are
combination of optics and instead leave to the user the possibility to
do it themselves
* Random query functions have been replaced by relevant optics
* Optics names have been homogenized and standardized
* Thinner optics have been defined to ease the use of values and minted
values
* Type-changing optics for fields with existential types have been
defined and the casting of such fields have been improved (fixes #488)

While I was homogenizing optics, I also homogenized the name of several
constructs which were inconsistent, such as `TxSkelOpts` and
`TxSkelLabels`. This PR seems huge because it changes many small things
in the codebase, but the main changes are actually focused in the
definition of the optics.